Move ConfigOptionsScreen to dedicated package
