+++ src/client/java/net/pneumono/pneumonocore/config_api/entries/CategoryTitleEntry.java
@@ -6,7 +6,7 @@ import net.minecraft.client.gui.Element;
 import net.minecraft.client.gui.Selectable;
 import net.minecraft.text.Text;
 import net.minecraft.util.Colors;
-import net.pneumono.pneumonocore.config_api.ConfigOptionsScreen;
+import net.pneumono.pneumonocore.config_api.screen.ConfigOptionsScreen;
 
 import java.util.List;
 import java.util.Objects;
